Split agent.py into executor and operator, handle safety check for CLI vs server
